+++ src/world.jl
@@ -132,8 +132,11 @@ end
     start::_GraphNode,
     add::Tuple{Vararg{Int}},
     remove::Tuple{Vararg{Int}},
+    add_mask::_Mask,
+    rem_mask::_Mask,
+    use_map::Union{_NoUseMap,_UseMap},
 )::UInt32
-    node = _find_node(world._graph, start, add, remove)
+    node = _find_node(world._graph, start, add, remove, add_mask, rem_mask, use_map)
 
     archetype = (node.archetype == typemax(UInt32)) ?
                 _create_archetype!(world, node) :
@@ -149,9 +152,12 @@ end
     remove::Tuple{Vararg{Int}},
     relations::Tuple{Vararg{Int}},
     targets::Tuple{Vararg{Entity}},
+    add_mask::_Mask,
+    rem_mask::_Mask,
+    use_map::Union{_NoUseMap,_UseMap},
 )::UInt32
     @inbounds old_arch = world._archetypes[old_table.archetype]
-    new_arch_index = _find_or_create_archetype!(world, old_arch.node, add, remove)
+    new_arch_index = _find_or_create_archetype!(world, old_arch.node, add, remove, add_mask, rem_mask, use_map)
     @inbounds new_arch = world._archetypes[new_arch_index]
 
     if !_has_relations(new_arch) && isempty(relations)
@@ -578,13 +584,22 @@ end
 
     _check_no_duplicates(add_types)
     _check_no_duplicates(rem_types)
+    _check_if_intersect(add_types, rem_types)
     _check_no_duplicates(rel_types)
     _check_relations(rel_types)
     _check_is_subset(rel_types, add_types)
 
-    add_ids = tuple([_component_id(W.parameters[1], T) for T in add_types]...)
-    rem_ids = tuple([_component_id(W.parameters[1], T) for T in rem_types]...)
-    rel_ids = tuple([_component_id(W.parameters[1], T) for T in rel_types]...)
+    CS = W.parameters[1]
+    add_ids = tuple([_component_id(CS, T) for T in add_types]...)
+    rem_ids = tuple([_component_id(CS, T) for T in rem_types]...)
+    rel_ids = tuple([_component_id(CS, T) for T in rel_types]...)
+
+    num_ids = length(add_ids) + length(rem_ids)
+    use_map = num_ids >= 4 ? _UseMap() : _NoUseMap()
+
+    M = max(1, cld(length(CS.parameters), 64))
+    add_mask = _Mask{M}(add_ids...)
+    rem_mask = _Mask{M}(rem_ids...)
 
     push!(exprs, :(index = world._entities[entity._id]))
     push!(exprs, :(old_table = world._tables[index.table]))
@@ -594,7 +609,7 @@ end
         :(
             new_table_index =
                 _find_or_create_table!(
-                    world, old_table, $add_ids, $rem_ids, $rel_ids, targets,
+                    world, old_table, $add_ids, $rem_ids, $rel_ids, targets, $add_mask, $rem_mask, $use_map,
                 )
         ),
     )
@@ -1079,11 +1094,33 @@ end
     _check_relations(rel_types)
     _check_is_subset(rel_types, types)
 
-    ids = tuple([_component_id(W.parameters[1], T) for T in types]...)
-    rel_ids = tuple([_component_id(W.parameters[1], T) for T in rel_types]...)
+    CS = W.parameters[1]
+    ids = tuple([_component_id(CS, T) for T in types]...)
+    rel_ids = tuple([_component_id(CS, T) for T in rel_types]...)
+    num_ids = length(ids)
+    use_map = num_ids >= 4 ? _UseMap() : _NoUseMap()
+
+    M = max(1, cld(length(CS.parameters), 64))
+    add_mask = _Mask{M}(ids...)
+    rem_mask = _Mask{M}()
 
     exprs = []
-    push!(exprs, :(table = _find_or_create_table!(world, world._tables[1], $ids, (), $rel_ids, targets)))
+    push!(
+        exprs,
+        :(
+            table = _find_or_create_table!(
+                world,
+                world._tables[1],
+                $ids,
+                (),
+                $rel_ids,
+                targets,
+                $add_mask,
+                $rem_mask,
+                $use_map,
+            )
+        ),
+    )
     push!(exprs, :(tmp = _create_entity!(world, table)))
     push!(exprs, :(entity = tmp[1]))
     push!(exprs, :(index = tmp[2]))
@@ -1273,11 +1310,33 @@ end
     _check_relations(rel_types)
     _check_is_subset(rel_types, types)
 
-    ids = tuple([_component_id(W.parameters[1], T) for T in types]...)
-    rel_ids = tuple([_component_id(W.parameters[1], T) for T in rel_types]...)
+    CS = W.parameters[1]
+    ids = tuple([_component_id(CS, T) for T in types]...)
+    rel_ids = tuple([_component_id(CS, T) for T in rel_types]...)
+    num_ids = length(ids)
+    use_map = num_ids >= 4 ? _UseMap() : _NoUseMap()
+
+    M = max(1, cld(length(CS.parameters), 64))
+    add_mask = _Mask{M}(ids...)
+    rem_mask = _Mask{M}()
 
     exprs = []
-    push!(exprs, :(table_idx = _find_or_create_table!(world, world._tables[1], $ids, (), $rel_ids, targets)))
+    push!(
+        exprs,
+        :(
+            table_idx = _find_or_create_table!(
+                world,
+                world._tables[1],
+                $ids,
+                (),
+                $rel_ids,
+                targets,
+                $add_mask,
+                $rem_mask,
+                $use_map,
+            )
+        ),
+    )
     push!(exprs, :(indices = _create_entities!(world, table_idx, n)))
     push!(exprs, :(table = world._tables[table_idx]))
 
@@ -1399,11 +1458,34 @@ end
     _check_relations(rel_types)
     _check_is_subset(rel_types, types)
 
-    ids = tuple([_component_id(W.parameters[1], T) for T in types]...)
-    rel_ids = tuple([_component_id(W.parameters[1], T) for T in rel_types]...)
+    CS = W.parameters[1]
+    ids = tuple([_component_id(CS, T) for T in types]...)
+    rel_ids = tuple([_component_id(CS, T) for T in rel_types]...)
+
+    num_ids = length(ids)
+    use_map = num_ids >= 4 ? _UseMap() : _NoUseMap()
+
+    M = max(1, cld(length(CS.parameters), 64))
+    add_mask = _Mask{M}(ids...)
+    rem_mask = _Mask{M}()
 
     exprs = []
-    push!(exprs, :(table_idx = _find_or_create_table!(world, world._tables[1], $ids, (), $rel_ids, targets)))
+    push!(
+        exprs,
+        :(
+            table_idx = _find_or_create_table!(
+                world,
+                world._tables[1],
+                $ids,
+                (),
+                $rel_ids,
+                targets,
+                $add_mask,
+                $rem_mask,
+                $use_map,
+            )
+        ),
+    )
     push!(exprs, :(indices = _create_entities!(world, table_idx, n)))
     push!(exprs, :(table = world._tables[table_idx]))
 
@@ -1547,15 +1629,24 @@ end
 
     _check_no_duplicates(add_types)
     _check_no_duplicates(rem_types)
+    _check_if_intersect(add_types, rem_types)
     _check_no_duplicates(rel_types)
     _check_relations(rel_types)
     _check_is_subset(rel_types, add_types)
 
     exprs = []
 
-    add_ids = tuple([_component_id(W.parameters[1], T) for T in add_types]...)
-    rem_ids = tuple([_component_id(W.parameters[1], T) for T in rem_types]...)
-    rel_ids = tuple([_component_id(W.parameters[1], T) for T in rel_types]...)
+    CS = W.parameters[1]
+    add_ids = tuple([_component_id(CS, T) for T in add_types]...)
+    rem_ids = tuple([_component_id(CS, T) for T in rem_types]...)
+    rel_ids = tuple([_component_id(CS, T) for T in rel_types]...)
+
+    num_ids = length(add_ids) + length(rem_ids)
+    use_map = num_ids >= 4 ? _UseMap() : _NoUseMap()
+
+    M = max(1, cld(length(CS.parameters), 64))
+    add_mask = _Mask{M}(add_ids...)
+    rem_mask = _Mask{M}(rem_ids...)
 
     push!(exprs, :(index = world._entities[entity._id]))
     push!(exprs, :(old_table = world._tables[index.table]))
@@ -1564,7 +1655,7 @@ end
         :(
             new_table_index =
                 _find_or_create_table!(
-                    world, old_table, $add_ids, $rem_ids, $rel_ids, targets,
+                    world, old_table, $add_ids, $rem_ids, $rel_ids, targets, $add_mask, $rem_mask, $use_map,
                 )
         ),
     )
@@ -1703,6 +1794,7 @@ end
     relations_vec = Expr(:vect, relations_expr...)
 
     M = max(1, cld(length(types), 64))
+    start_mask = _Mask{M}()
     return quote
         registry = _ComponentRegistry()
         ids = $id_tuple
@@ -1719,7 +1811,7 @@ end
             targets,
             $storage_tuple,
             $relations_vec,
-            [_Archetype(UInt32(1), first(graph.nodes)[2], _TableIDs(zero_table))],
+            [_Archetype(UInt32(1), graph.nodes[$start_mask], _TableIDs(zero_table))],
             Vector{_Archetype{$(M)}}(),
             [zero_table],
             _ComponentIndex{$(M)}($(length(types))),